+++ geojson/src/jvmMain/kotlin/io/data2viz/geojson/toGeoJsonObject.kt
@@ -59,9 +59,9 @@ fun Collection<Collection<Collection<LngLatAlt>>>.toSurfaces(): Array<Lines> = m
 
 actual class FeatureProperties {
 	var properties: Map<String, Any?> = mapOf()
-    actual fun stringProp(name: String): String  	= properties[name] as String
-	actual fun intProp(name: String): Int  			= properties[name] as Int
-	actual fun booleanProp(name: String): Boolean  	= properties[name] as Boolean
+    actual fun stringProperty(name: String): String  	= properties[name] as String
+	actual fun intProperty(name: String): Int  			= properties[name] as Int
+	actual fun booleanProperty(name: String): Boolean  	= properties[name] as Boolean
 }
 
 actual fun <T> String.toFeaturesAndProperties(extractFunction: FeatureProperties.() -> T): List<Pair<Feature, T>> {