+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.1-RC2
+
+> Published 23 Jan 2020
+* rename `intProp`, `booleanProp`, `stringProp` to 
+`intProperty`, `booleanProperty`, `stringProperty`
+
 ## 0.6.1-RC1
 
 > Published 22 Jan 2020